Vacation
I was the hole vacation not at home and didnt have any internet connection.
So i have made everything and upload it to github just now because i could not do it earlier.
I made a character component that uses states and command.
I did it this way so you can easily change up things and so a player can controle a fygar or pooga.
I added a level component and i am thinking of a level manager so you can add the levels to a manager.
I started afew little things but nothing is finished or even close to.